Uploaded 5b papyrus project and image
Even though the name says 5a, this is 5b. I used signatures that were used in my 5a thus I had to use the same model for both diagrams.
Papyrus was also extremely buggy for this diagram type for an unknown reason. This is a big reason why this diagram is a bit messy as I was unable to resize or move a lot of the elements randomly. Some screenshots of errors encountered are also updated as evidence. These errors prevented me from doing things like moving elements, changing names or resizing them.